处理清查分组表单

+++ assess-app/src/main/java/com/copower/pmcc/assess/dal/basis/mapper/SurveyAssetInfoGroupMapper.java
@@ -6,7 +6,7 @@ import java.util.List;
 import org.apache.ibatis.annotations.Param;
 
 public interface SurveyAssetInfoGroupMapper {
-    int countByExample(SurveyAssetInfoGroupExample example);
+    long countByExample(SurveyAssetInfoGroupExample example);
 
     int deleteByExample(SurveyAssetInfoGroupExample example);
